Switch to Mahiru + Napi, Refactor Files & New Color Scheme
